Update to not error with latest chocolatey version
The VBoxManage install command was returning the percentage progress via
stderr, which caused chocolatey to assume failure (even though it did
install the ExtensionPack successfully). This fix captures the
stdout/stderr, then either displays it on success or uses it as an
exception message on failure, which causes chocolatey to correctly
determine success or failure of the VB ExtensionPack installation.